style: apply rustfmt + clippy fix to dugite-ledger/dugite-node (drive-by)
Recent perf(blockfetch) commits on main left rustfmt drift and a new
`clippy::unnecessary_cast` lint. Without these tiny drive-by fixes,
this PR's `build-and-test` job is blocked by failures inherited from
main's HEAD.

- dugite-ledger/state/apply.rs:689,1125 — rustfmt rewrap
- dugite-node/connection_lifecycle.rs:1264 — rustfmt one-line let
- dugite-node/connection_lifecycle.rs:1279 — drop unnecessary `as u64`
  cast (`std::collections::hash_map::DefaultHasher::finish()` is
  already `u64`)
- dugite-node/mod.rs — one rustfmt rewrap

+++ crates/dugite-node/src/node/connection_lifecycle.rs
@@ -1276,7 +1276,7 @@ impl ConnectionLifecycleManager {
                 let phase_offset = {
                     let mut h = std::collections::hash_map::DefaultHasher::new();
                     addr.hash(&mut h);
-                    std::time::Duration::from_millis((h.finish() % 200) as u64)
+                    std::time::Duration::from_millis(h.finish() % 200)
                 };
                 let mut poll_ticker = tokio::time::interval_at(
                     tokio::time::Instant::now() + phase_offset,
@@ -2532,8 +2532,7 @@ mod tests {
         for h in &[h2, h4] {
             fetched_hashes.insert(*h);
         }
-        let drained =
-            select_headers_to_fetch(&pending, |h| known.contains(h), &fetched_hashes);
+        let drained = select_headers_to_fetch(&pending, |h| known.contains(h), &fetched_hashes);
         assert!(
             drained.is_empty(),
             "all headers delivered — no more work to schedule"
